scaffold: Adding non-wizard workflow path
